+++ packages/ngx-layout/src/lib/footer/components/subfooter/subfooter.component.ts
@@ -1,4 +1,4 @@
-import {ChangeDetectionStrategy, Component} from '@angular/core';
+import {ChangeDetectionStrategy, Component, Input} from '@angular/core';
 
 @Component({
   selector: 'aui-subfooter',
@@ -6,6 +6,8 @@ import {ChangeDetectionStrategy, Component} from '@angular/core';
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class SubFooterComponent {
+  @Input() public ariaGoUpLabel = 'Ga terug naar boven';
+
   public goToTop() {
     window.scrollTo(0, 0);
   }