+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## Unreleased
 
+#### Fixed
+ - `layout` Added missing ARIA label for the scroll to top button in the footer
+
+#### Changed
+ - `table` Removed `short-unique-id` dependency as it was giving optimization bailouts warnings
+
 
 ## [4.5.1] - 2020-10-14